2.10 Update2.0
The logic of switching between the week and the month is filled, and the self-switching of the month and the week still needs to be filled. At the same time, some logic modifications and supplements are included. Some adjustments have also been made on the UI, but the UI part has not yet been divided.